removes SearchService and tidies RestaurantService.js
