+++ client/screens/TabHomeScreen.js
@@ -1,7 +1,7 @@
 import React, {useState, useEffect} from 'react';
 import {TouchableOpacity,StyleSheet,View,Text,SafeAreaView, ScrollView, Image} from 'react-native';
 import FilteredRestaurants from '../containers/FilteredRestaurants';
-import { getRestaurants } from '../services/SearchService'
+import { getRestaurants } from '../services/RestaurantService'
 import { getRestaurantById } from '../services/RestaurantService';
 import logo from './scranscanner-icon-white.png'
 import moment from 'moment';
